requirements/base.txt
django==1.11
django-autoslug==1.9.3
Pillow==4.3.0
django-crispy-forms==1.7.2
xlwt==1.3.0
django-registration-redux==2.6

